refactor(core): split the flattener into a document and stages
Replace the 1054-line class with SvgDocument (element tree, shape cache
and tree sync, traversal with transform/clip/attribute context, viewBox,
url(#id) lookup, serialization) and one function per stage under
flatten/stages, ordered in flatten/stages.ts. Stage bodies are moved
verbatim. Element-level helpers live in flatten/element.ts, inline
regexes become named constants in utils/svgPatterns.ts, and dead
exports, the safeRemove alias and the parseSvgTransform alias go.
The per-file port attribution headers are replaced by the LICENSES.md
entry, which now ships with the package.
