+++ packages/react-native-nano-icons/__tests__/flatten-primitives.unit.test.ts
@@ -1,12 +1,11 @@
 /** @jest-environment node */
 
-// Expected values in the fixture were generated by running the original
-// picosvg 0.22.3 python source (svg_transform / arc_to_cubic / svg_path_iter /
-// svg_types) on the same inputs. The TS port must reproduce them exactly.
+// Expected values in the fixture were generated by the reference
+// implementation on the same inputs and must be reproduced exactly.
 
 import expected from './fixtures/flatten-primitives-expected.json';
 import { ntos, pt, pythonRound } from '../src/core/flatten/geometry';
-import { Affine2D, parseSvgTransform } from '../src/core/flatten/transform';
+import { Affine2D } from '../src/core/flatten/transform';
 import { arcToCubic } from '../src/core/flatten/arcs';
 import {
   absoluteMoveto,
@@ -38,24 +37,24 @@ function expectNumbersClose(
   }
 }
 
-describe('parseSvgTransform', () => {
+describe('Affine2D.fromString', () => {
   for (const [raw, values] of Object.entries(expected.transforms)) {
     test(`"${raw}"`, () => {
-      expectNumbersClose(parseSvgTransform(raw).values(), values);
+      expectNumbersClose(Affine2D.fromString(raw).values(), values);
     });
   }
 
   test('compose_ltr', () => {
     const composed = Affine2D.composeLtr([
-      parseSvgTransform('rotate(30)'),
-      parseSvgTransform('translate(10 20)'),
-      parseSvgTransform('scale(2)'),
+      Affine2D.fromString('rotate(30)'),
+      Affine2D.fromString('translate(10 20)'),
+      Affine2D.fromString('scale(2)'),
     ]);
     expect(composed.values()).toEqual(expected.compose_ltr);
   });
 
   test('inverse', () => {
-    const inv = parseSvgTransform('translate(1,2) rotate(30) scale(2,3)')
+    const inv = Affine2D.fromString('translate(1,2) rotate(30) scale(2,3)')
       .inverse()
       .values();
     expect(inv).toEqual(expected.inverse);
@@ -73,10 +72,10 @@ describe('parseSvgTransform', () => {
   }
 
   test('toString', () => {
-    expect(parseSvgTransform('translate(10 20)').toString()).toBe(
+    expect(Affine2D.fromString('translate(10 20)').toString()).toBe(
       expected.tostring['translate(10 20)']
     );
-    expect(parseSvgTransform('rotate(45)').toString()).toBe(
+    expect(Affine2D.fromString('rotate(45)').toString()).toBe(
       expected.tostring['rotate(45)']
     );
     expect(Affine2D.identity().toString()).toBe(expected.tostring.identity);
