+++ packages/react-native-nano-icons/src/core/flatten/stages/opacity.ts
@@ -1,0 +1,6 @@
+import type { SvgDocument } from '../document';
+import { normalizeOpacity } from '../shape';
+
+export function normalizeOpacities(doc: SvgDocument): void {
+  doc.forEachShape((s) => normalizeOpacity(s));
+}
